pyproject.toml
[tool.poetry]
name = "spimex-app-fastapi"
version = "0.1.0"
description = ""
authors = ["PavelShaura <gkdrdfffe@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
black = "^24.4.2"
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
pydantic-settings = "^2.3.4"
alembic = "^1.13.2"
redis = "^5.0.7"
aiohttp = "^3.9.5"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
xlrd = "^2.0.1"
celery = {extras = ["redis"], version = "^5.4.0"}
fastapi-cache2 = "0.1.9"
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
pytest-dotenv = "^0.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "./app", "app",
]

env_files = [
  ".test.env"
]

asyncio_mode="auto"
